Allow providing request config file to request and read-latest tasks
