--- samples/common.js
@@ -1,23 +1,0 @@
-const Location = {
-  Inline: 0,
-}
-
-const CodeLanguage = {
-  JavaScript: 0,
-}
-
-const ReturnType = {
-  uint: "uint256",
-  uint256: "uint256",
-  int: "int256",
-  int256: "int256",
-  string: "string",
-  bytes: "Buffer",
-  Buffer: "Buffer",
-}
-
-module.exports = {
-  Location,
-  CodeLanguage,
-  ReturnType,
-}
